+++ src/components/Sections/Testimonials.tsx
@@ -127,7 +127,7 @@ function TestimonialCard({testimonial: {text, name, image, alt}, isActive}: {tes
             {image ? (
                 <div className="relative h-14 w-14 shrink-0 sm:h-16 sm:w-16">
                     <QuoteIcon className="absolute -left-2 -top-2 h-4 w-4 stroke-black text-white" />
-                    <Image alt={alt} className="rounded-full" height={64} src={image} width={64} />
+                    <Image alt={alt} className="rounded-full" height={64} loading="lazy" src={image} width={64} />
                 </div>
             ) : (
                 <QuoteIcon className="h-5 w-5 shrink-0 text-white sm:h-8 sm:w-8" />